Completed 2018 mapping

+++ src/backend/FantasyHOF.Domain/Types/TeamSeasonStats.cs
@@ -8,8 +8,8 @@
         public required int Wins { get; set; }
         public required int Losses { get; set; }
         public required int Ties { get; set; }
-        public required double WinPercentage { get; set; }
-        public required float PointsAgainst { get; set; }
-        public required float PointsFor { get; set; }
+        public required decimal WinPercentage { get; set; }
+        public required decimal PointsAgainst { get; set; }
+        public required decimal PointsFor { get; set; }
     }
 }
