+++ src/backend/FantasyHOF.ESPN/Types/Models/ESPNMatchupTeam.cs
@@ -3,7 +3,7 @@
     public class ESPNMatchupTeam
     {
         public required int TeamId { get; set; }
-        public required float TotalPoints { get; set; }
+        public required decimal TotalPoints { get; set; }
         public required ESPNRoster rosterForCurrentScoringPeriod { get; set; }
     }
 }
